Make Page() a method of Outline
I didn't realize I already had a handy reference to *pdf.Reader
present in Outline (by way of Value).  This means Page() can be a
method of Outline, rather than a method of *pdf.Reader, which seems to
me a more logical place for it.  Best to change this now, before
anyone else might be using this code.